refactor: extract sdk error func into helpers file, rename import aliases
